# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Finance MVP API
  version: 0.1.0
  summary: Canonical OpenAPI contract for the mounted closed-MVP runtime surface.
  license:
    name: Proprietary/Internal
    url: https://example.invalid/proprietary-internal-license
  description: |
    This document is the source of truth for the mounted MVP wire routes, DTO field names,
    enum values, error codes, pagination envelopes, and auth requirements.

    MVP boundaries are intentionally narrow: financial data is manual-entry only,
    `sourceType` accepts only `manual`, and transfers are same-scope only. Safe
    screenshot OCR capture may create pending drafts from sanitized structured
    candidates, but raw image bytes and OCR text are rejected and confirmed
    transactions remain `sourceType=manual`. There is no real import, parser,
    file storage, bank API, automated ingestion, broker credential, external
    credential, card, IBAN, raw statement, or device-token endpoint.

    Privacy invariants:
    - personal data is owner-only;
    - shared data is visible only to active members of the same Household;
    - reports support `personal`, `shared_family_report`, and `combined_viewer_overview`;
    - every list, report, search, autocomplete, cache, count, and aggregation
      filters visible rows before any aggregation or pagination;
    - responses never expose hidden counts, filtered-out counts, hidden facets,
      or diagnostics that confirm inaccessible resources.
servers:
...
security:
  - pwaCookieAuth: []
    csrfHeader: []
  - androidBearerAuth: []
paths:
...
  /categories:
    get:
      tags: [Categories]
      operationId: listCategories
      summary: List/search visible categories.
      description: Personal categories are owner-only; household categories require active membership. Usage counts for hidden personal activity are never exposed.
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/CategoryScopeFilter'
        - $ref: '#/components/parameters/CategoryTypeFilter'
        - $ref: '#/components/parameters/HouseholdIdFilter'
        - $ref: '#/components/parameters/StatusFilter'
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: Visible categories.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryPageEnvelope'
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Categories]
      operationId: createCategory
      summary: Create personal or household category.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreateRequest'
      responses:
        '201':
          description: Created category.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryEnvelope'
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /categories/autocomplete:
    get:
      tags: [Categories]
      operationId: autocompleteCategories
      summary: Autocomplete visible categories.
      parameters:
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/CategoryScopeFilter'
        - $ref: '#/components/parameters/CategoryTypeFilter'
        - $ref: '#/components/parameters/HouseholdIdFilter'
      responses:
        '200':
          description: Category suggestions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryAutocompleteListEnvelope'
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /categories/{categoryId}:
    get:
      tags: [Categories]
      operationId: getCategory
      summary: Get visible category detail.
      parameters:
        - $ref: '#/components/parameters/CategoryId'
      responses:
        '200':
          description: Category detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryEnvelope'
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
    patch:
      tags: [Categories]
      operationId: updateCategory
      summary: Update category mutable fields.
      parameters:
        - $ref: '#/components/parameters/CategoryId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryUpdateRequest'
      responses:
        '200':
          description: Updated category.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryEnvelope'
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags: [Categories]
      operationId: deleteCategory
      summary: Soft delete category if history remains coherent.
      parameters:
        - $ref: '#/components/parameters/CategoryId'
      responses:
        '204':
          description: Category soft-deleted.
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /categories/{categoryId}/archive:
    post:
      tags: [Categories]
      operationId: archiveCategory
      summary: Archive category.
      parameters:
        - $ref: '#/components/parameters/CategoryId'
      responses:
        '200':
          description: Archived category.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryEnvelope'
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /categories/{categoryId}/restore:
    post:
      tags: [Categories]
      operationId: restoreCategory
      summary: Restore category.
      parameters:
        - $ref: '#/components/parameters/CategoryId'
      responses:
        '200':
          description: Restored category.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryEnvelope'
        '4XX':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    pwaCookieAuth:
      type: apiKey
      in: cookie
      name: __Host-finance_session
      description: HttpOnly Secure SameSite cookie used by the PWA.
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Required for PWA cookie-authenticated state-changing requests; bound to the authenticated session and rotated on login/logout/reset.
    androidBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Opaque bearer access token for Android; refresh token storage and rotation are platform secure-storage concerns.
  responses:
    Error:
      description: Canonical error envelope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  parameters:
...
    CategoryId:
      name: categoryId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/ResourceId' }
...
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
        minLength: 1
    Sort:
      name: sort
      in: query
      required: false
      schema:
        type: string
        minLength: 1
      description: Endpoint-specific allowlisted sort value.
    SearchQuery:
      name: q
      in: query
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 200
      description: Search runs only after visible-scope filtering and is not logged as raw payload.
...
    HouseholdIdFilter:
      name: householdId
      in: query
      required: false
      schema: { $ref: '#/components/schemas/ResourceId' }
      description: Optional scope narrowing only; inaccessible ids use neutral behavior.
...
    CategoryScopeFilter:
      name: scope
      in: query
      required: false
      schema: { $ref: '#/components/schemas/CategoryScope' }
    CategoryTypeFilter:
      name: type
      in: query
      required: false
      schema: { $ref: '#/components/schemas/CategoryType' }
...
    StatusFilter:
      name: status
      in: query
      required: false
      schema: { $ref: '#/components/schemas/RecordStatus' }
...
  schemas:
    ResourceId:
      type: string
      minLength: 1
      maxLength: 128
      pattern: '^[A-Za-z0-9_:-]+$'
      examples: [usr_123]
    DateTime:
      type: string
      format: date-time
      examples: ['2026-05-17T10:00:00Z']
...
    CategoryScope:
      type: string
      enum: [personal, household]
      description: Category scope uses `household`, not `shared`.
    CategoryType:
      type: string
      enum: [income, expense]
...
    RecordStatus:
      type: string
      enum: [active, archived, deleted]
...
    ErrorCode:
      type: string
      enum:
        - UNAUTHENTICATED
        - SESSION_EXPIRED
        - SESSION_REVOKED
        - RESOURCE_NOT_FOUND_OR_NOT_ACCESSIBLE
        - REFERENCED_RESOURCE_NOT_FOUND_OR_NOT_ACCESSIBLE
        - ACTION_NOT_ALLOWED
        - MEMBERSHIP_NOT_ACTIVE
        - VALIDATION_FAILED
        - INVALID_ENUM_VALUE
        - INVALID_DATE_RANGE
        - INVALID_CURRENCY
        - ACCOUNT_OWNERSHIP_IMMUTABLE
        - CATEGORY_SCOPE_MISMATCH
        - TRANSACTION_ACCOUNT_MISMATCH
        - ARCHIVED_RECORD_NOT_MUTABLE
        - CONFLICTING_UPDATE
        - ACCOUNT_CURRENCY_IMMUTABLE_AFTER_TRANSACTIONS
        - TRANSFER_SCOPE_NOT_SUPPORTED
        - TRANSFER_COUNTERPARTY_REQUIRED
        - TRANSFER_COUNTERPARTY_NOT_ACCESSIBLE
        - TOO_MANY_REQUESTS
        - RATE_LIMITED
        - LOGIN_RATE_LIMITED
        - RESET_RATE_LIMITED
        - INVITE_RATE_LIMITED
        - UPLOAD_TOO_LARGE
        - IMAGE_TOO_LARGE
        - UNSUPPORTED_MEDIA_TYPE
        - INVALID_IMAGE
        - OCR_DISABLED
        - OCR_ENGINE_UNAVAILABLE
        - OCR_TIMEOUT
        - INVITE_NOT_FOUND_OR_NOT_ACCESSIBLE
        - INVITE_EXPIRED
        - INVITE_ALREADY_USED
        - INVITE_REVOKED
        - HOUSEHOLD_MEMBER_LIMIT_REACHED
        - RESET_TOKEN_NOT_FOUND_OR_EXPIRED
        - RESET_TOKEN_ALREADY_USED
        - PASSWORD_POLICY_FAILED
        - IDEMPOTENCY_KEY_REUSED
        - UNSUPPORTED_ENTITY_TYPE
        - UNSUPPORTED_OPERATION
        - UNSUPPORTED_TRANSACTION_TYPE
        - PAYLOAD_REQUIRED
        - BASE_VERSION_REQUIRED
        - ENTITY_ALREADY_EXISTS
    ErrorDetail:
      type: object
      description: Field-level details for caller-supplied invalid fields only; never includes hidden object names, amounts, emails, tokens, SQL, stack traces, or raw payloads.
      additionalProperties: false
      properties:
        field:
          type: string
        reason:
          type: string
        allowedValues:
          type: array
          items:
            type: string
    ErrorDto:
      type: object
      additionalProperties: false
      required: [code, message, requestId]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: Safe generic user-facing message.
        requestId:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorDto'
    PageInfo:
      type: object
      additionalProperties: false
      required: [limit, hasMore]
      properties:
        limit:
          type: integer
          minimum: 1
          maximum: 100
        nextCursor:
          type: [string, 'null']
          description: Opaque cursor scoped to visible filtered rows only.
        hasMore:
          type: boolean
      description: Pagination metadata never includes totalCount or hidden counts.
...
    CategoryDto:
      type: object
      additionalProperties: false
      required: [id, name, type, scope, status, createdByUserId, createdAt, updatedAt, version]
      properties:
        id:
          $ref: '#/components/schemas/ResourceId'
        name:
          type: string
          maxLength: 120
        type:
          $ref: '#/components/schemas/CategoryType'
        scope:
          $ref: '#/components/schemas/CategoryScope'
        ownerUserId:
          anyOf:
            - $ref: '#/components/schemas/ResourceId'
            - type: 'null'
        householdId:
          anyOf:
            - $ref: '#/components/schemas/ResourceId'
            - type: 'null'
        iconKey:
          type: [string, 'null']
          maxLength: 80
        color:
          type: [string, 'null']
          pattern: '^#[0-9A-Fa-f]{6}$'
        status:
          $ref: '#/components/schemas/RecordStatus'
        createdByUserId:
          $ref: '#/components/schemas/ResourceId'
        createdAt:
          $ref: '#/components/schemas/DateTime'
        updatedAt:
          $ref: '#/components/schemas/DateTime'
        archivedAt:
          anyOf:
            - $ref: '#/components/schemas/DateTime'
            - type: 'null'
        deletedAt:
          anyOf:
            - $ref: '#/components/schemas/DateTime'
            - type: 'null'
        version:
          type: integer
          minimum: 1
    CategoryCreateRequest:
      type: object
      additionalProperties: false
      required: [name, type, scope]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        type:
          $ref: '#/components/schemas/CategoryType'
        scope:
          $ref: '#/components/schemas/CategoryScope'
        householdId:
          $ref: '#/components/schemas/ResourceId'
          description: Required for household categories.
        iconKey:
          type: string
          maxLength: 80
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
    CategoryUpdateRequest:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        iconKey:
          type: string
          maxLength: 80
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
        status:
          $ref: '#/components/schemas/RecordStatus'
        version:
          type: integer
          minimum: 1
      description: scope, ownerUserId, and householdId are immutable in MVP.
    CategoryAutocompleteDto:
      type: object
      additionalProperties: false
      required: [id, name, type, scope]
      properties:
        id:
          $ref: '#/components/schemas/ResourceId'
        name:
          type: string
        type:
          $ref: '#/components/schemas/CategoryType'
        scope:
          $ref: '#/components/schemas/CategoryScope'
        householdId:
          anyOf:
            - $ref: '#/components/schemas/ResourceId'
            - type: 'null'
        iconKey:
          type: [string, 'null']
        color:
          type: [string, 'null']
...
    CategoryEnvelope:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/CategoryDto'
    CategoryPageEnvelope:
      type: object
      additionalProperties: false
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CategoryDto'
        page:
          $ref: '#/components/schemas/PageInfo'
    CategoryAutocompleteListEnvelope:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CategoryAutocompleteDto'
...